simulation/requirements.txt
fastapi
uvicorn
pydantic
requests
loguru
tqdm
numpy
sionna-rt